Align report policy and clean helper imports
